[public-api] Restructure packages to better align with other components
