+++ components/public-api-server/main.go
@@ -5,45 +5,14 @@
 package main
 
 import (
-	"github.com/gitpod-io/gitpod/common-go/baseserver"
 	"github.com/gitpod-io/gitpod/common-go/log"
-	"github.com/gitpod-io/gitpod/public-api-server/middleware"
-	"github.com/gitpod-io/gitpod/public-api-server/pkg/apiv1"
-	v1 "github.com/gitpod-io/gitpod/public-api/v1"
-	"net/http"
+	"github.com/gitpod-io/gitpod/public-api-server/pkg/server"
 )
 
 func main() {
 	logger := log.New()
-	srv, err := baseserver.New("public_api_server",
-		baseserver.WithLogger(logger),
-		baseserver.WithHTTPPort(9000),
-		baseserver.WithGRPCPort(9001),
-	)
-	if err != nil {
-		logger.WithError(err).Fatal("Failed to initialize public api server.")
+
+	if err := server.Start(logger); err != nil {
+		logger.WithError(err).Fatal("Server errored.")
 	}
-
-	if err := register(srv); err != nil {
-		logger.WithError(err).Fatal("Failed to register services.")
-	}
-
-	if listenErr := srv.ListenAndServe(); listenErr != nil {
-		logger.WithError(listenErr).Fatal("Failed to serve public api server")
-	}
-}
-
-func register(srv *baseserver.Server) error {
-	logger := log.New()
-	m := middleware.NewLoggingMiddleware(logger)
-	srv.HTTPMux().Handle("/", m(http.HandlerFunc(HelloWorldHandler)))
-
-	v1.RegisterWorkspacesServiceServer(srv.GRPC(), apiv1.NewWorkspaceService())
-	v1.RegisterPrebuildsServiceServer(srv.GRPC(), v1.UnimplementedPrebuildsServiceServer{})
-
-	return nil
-}
-
-func HelloWorldHandler(w http.ResponseWriter, _ *http.Request) {
-	w.Write([]byte(`hello world`))
 }
