+++ components/public-api-server/pkg/server/integration_test.go
@@ -2,7 +2,7 @@
 // Licensed under the GNU Affero General Public License (AGPL).
 // See License-AGPL.txt in the project root for license information.
 
-package main
+package server
 
 import (
 	"context"
